refactor: address PR outcome analytics review feedback
